+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,6 +1,6 @@
 object Versions {
   const val compileSdkVersion = 30
-  const val minSdkVersion = 15
+  const val minSdkVersion = 21
   const val targetSdkVersion = 30
 
   const val kotlinVersion = "1.3.72"
