+++ magellan-library/src/test/java/com/wealthfront/magellan/navigation/LinearNavigatorTest.kt
@@ -163,6 +163,7 @@ class LinearNavigatorTest {
 
     linearNavigator.goBack()
     linearNavigator.goBack()
+    linearNavigator.pause(context)
     linearNavigator.destroy(context)
 
     assertThat(linearNavigator.backStack.size).isEqualTo(0)
