+++ magellan-sample-advanced/build.gradle
@@ -8,7 +8,7 @@ android {
 
   defaultConfig {
     applicationId "com.wealthfront.magellan.sample"
-    minSdkVersion 21
+    minSdkVersion Versions.minSdkVersion
     targetSdkVersion Versions.targetSdkVersion
     versionCode 1
     versionName "1.0"
